pyproject.toml
[tool.poetry]
name = "sde_bbdm"
version = "1.1b1"
description = "Score-Based Image-to-Image Brownian Bridge (v1.1 Beta 1)"
authors = []
packages = [
    { include = "sde_bbdm" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.2"
torchmanager = "^1.2"
torchmanager-diffusion = "^1.1b4"
einops = "^0.6.1"
lpips = "^0.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
